+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## [UNRELEASED]
 
+### Added
+- 'casing': 'title' #1277
+
 ### Updates
 - performance updates
 	- getmasklength